+++ lib/theme_data.dart
@@ -1,43 +1,77 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:flex_color_scheme/flex_color_scheme.dart';
 
 // Nice to get colors: https://icolorpalette.com/color/101920
 
+// Scheme builder: https://rydmike.com/flexcolorscheme/themesplayground-v8-0/
+
 const Color checkMKIconColor = Color(0xFF13D389);
 
-ThemeData buildLightTheme() {
-  final base = ThemeData(
-    useMaterial3: true,
+/// The [AppTheme] defines light and dark themes for the app.
+///
+/// Theme setup for FlexColorScheme package v8.
+/// Use same major flex_color_scheme package version. If you use a
+/// lower minor version, some properties may not be supported.
+///
+/// Use in [MaterialApp] like this:
+///
+/// MaterialApp(
+///  theme: AppTheme.light,
+///  darkTheme: AppTheme.dark,
+///  :
+/// );
+sealed class AppTheme {
+  // The defined light theme.
+  static ThemeData lightBase = FlexThemeData.light(
+    scheme: FlexScheme.greenM3,
+    subThemesData: const FlexSubThemesData(
+      interactionEffects: true,
+      tintedDisabledControls: true,
+      useM2StyleDividerInM3: true,
+      inputDecoratorIsFilled: true,
+      inputDecoratorBorderType: FlexInputBorderType.outline,
+      alignedDropdown: true,
+      navigationRailUseIndicator: true,
+      navigationRailLabelType: NavigationRailLabelType.all,
+    ),
+    keyColors: const FlexKeyColors(),
+    visualDensity: FlexColorScheme.comfortablePlatformDensity,
+    cupertinoOverrideTheme: const CupertinoThemeData(applyThemeToAll: true),
+  );
+  static ThemeData light = lightBase.copyWith(
     appBarTheme: AppBarTheme(
       color: checkMKIconColor,
       iconTheme: IconThemeData(color: Colors.white),
     ),
-    primaryTextTheme: Typography.material2021().white,
-    textTheme: Typography.material2021().black,
-    brightness: Brightness.light,
-    colorScheme: ColorScheme.fromSeed(seedColor: checkMKIconColor),
+    textTheme: GoogleFonts.latoTextTheme(lightBase.textTheme),
   );
 
-  return base.copyWith(
-    textTheme: GoogleFonts.latoTextTheme(base.textTheme),
+  // The defined dark theme.
+  static ThemeData darkBase = FlexThemeData.dark(
+    scheme: FlexScheme.greenM3,
+    subThemesData: const FlexSubThemesData(
+      interactionEffects: true,
+      tintedDisabledControls: true,
+      blendOnColors: true,
+      useM2StyleDividerInM3: true,
+      inputDecoratorIsFilled: true,
+      inputDecoratorBorderType: FlexInputBorderType.outline,
+      alignedDropdown: true,
+      navigationRailUseIndicator: true,
+      navigationRailLabelType: NavigationRailLabelType.all,
+    ),
+    keyColors: const FlexKeyColors(),
+    visualDensity: FlexColorScheme.comfortablePlatformDensity,
+    cupertinoOverrideTheme: const CupertinoThemeData(applyThemeToAll: true),
   );
-}
-
-ThemeData buildDarkTheme() {
-  final base = ThemeData(
-    useMaterial3: true,
+  static ThemeData dark = darkBase.copyWith(
     appBarTheme: AppBarTheme(
       color: checkMKIconColor,
       iconTheme: IconThemeData(color: Colors.white),
     ),
-    primaryTextTheme: Typography.material2021().white,
-    textTheme: Typography.material2021().white,
-    brightness: Brightness.dark,
-    colorScheme: ColorScheme.fromSeed(
-        seedColor: checkMKIconColor, brightness: Brightness.dark),
+    textTheme: GoogleFonts.latoTextTheme(darkBase.textTheme),
   );
 
-  return base.copyWith(
-    textTheme: GoogleFonts.latoTextTheme(base.textTheme),
-  );
 }